pyproject.toml
[tool.poetry]
name = "hedging-dataflow"
version = "0.1.0"
description = ""
authors = ["cuckflong <johnathanclong13@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.8"
prefect = "^2.4.2"
ctrader-open-api = "^0.9.1"
psycopg2 = "^2.9.3"
substrate-interface = "^1.3.1"
s3fs = "^2022.8.2"
pyOpenSSL = "^22.1.0"
service-identity = "^21.1.0"
psycopg2-binary = "^2.9.3"
python-binance = "^1.0.16"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
